+++ src/Controller/ClientUserController.php
@@ -8,23 +8,16 @@
 
 namespace App\Controller;
 
-use App\Entity\Client;
 use App\Entity\User;
-use App\Form\Type\UserRegistrationType;
 use App\Security\JwtTokenAuthenticator;
 use App\Service\EntityManager\UserManager;
 use FOS\RestBundle\Controller\FOSRestController;
 use FOS\RestBundle\Controller\Annotations as Rest;
-use FOS\RestBundle\Request\ParamFetcherInterface;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
-use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
-use Symfony\Component\Validator\ConstraintViolationList;
 
 class ClientUserController extends FOSRestController
 {
