+++ src/Service/EntityManager/UserManager.php
@@ -11,13 +11,11 @@ namespace App\Service\EntityManager;
 
 use App\Entity\Client;
 use App\Entity\User;
-use App\Form\Type\UserRegistrationType;
 use App\Security\JwtTokenAuthenticator;
+use App\Service\Helper\FormHelper;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
-use Symfony\Component\Form\FormInterface;
 
 
 
@@ -25,26 +23,28 @@ class UserManager
 {
     private $em;
     private $repository;
-    private $formFactory;
     private $passwordEncoder;
     private $authenticator;
+    private $formHelper;
 
     public function __construct(
         UserPasswordEncoderInterface $passwordEncoder,
-        FormFactoryInterface $formFactory,
         EntityManagerInterface $em,
-        JwtTokenAuthenticator $authenticator
-    ) {
+        JwtTokenAuthenticator $authenticator,
+        FormHelper $formHelper
+    )
+    {
         $this->em = $em;
         $this->repository = $em->getRepository(User::class);
-        $this->formFactory = $formFactory;
         $this->passwordEncoder = $passwordEncoder;
         $this->authenticator = $authenticator;
+        $this->formHelper = $formHelper;
     }
 
     /**
      * @param Request $request
      * @param $id
+     * @return User|null
      * @throws \Doctrine\ORM\NonUniqueResultException
      */
     public function getOneUserByClient(Request $request, $id)
@@ -54,6 +54,10 @@ class UserManager
         return $searchedUser;
     }
 
+    /**
+     * @param Request $request
+     * @return mixed
+     */
     public function getAllUsersByClient(Request $request)
     {
         $clientId = $this->getClientId($request);
@@ -61,6 +65,10 @@ class UserManager
         return $clientUsers;
     }
 
+    /**
+     * @param Request $request
+     * @return Client|null|object
+     */
     public function getClient(Request $request)
     {
         $clientId = $this->getClientId($request);
@@ -68,17 +76,25 @@ class UserManager
         return $client;
     }
 
+    /**
+     * @param Request $request
+     * @return mixed
+     */
     public function getClientId(Request $request)
     {
         $userData = $this->authenticator->getUserUtils($request);
         return $userData['client'];
     }
 
+    /**
+     * @param Request $request
+     * @return User|array
+     */
     public function registerUser(Request $request)
     {
         $client = $this->getClient($request);
         $user = new User();
-        $form = $this->createForm(UserRegistrationType::class, $user);
+        $form = $this->formHelper->createUserRegistrationForm($user);
         $form->submit($request->request->all(), true);
         if ($form->isSubmitted() && $form->isValid()) {
             $password = $this->passwordEncoder->encodePassword($user, $user->getPlainPassword());
@@ -89,46 +105,12 @@ class UserManager
             $this->em->flush();
             return $user;
         }
-
-        $errors = $this->getErrorsFromForm($form);
-        $data = [
-            'type' => 'validation_error',
-            'title' => 'There was a validation error',
-            'errors' => $errors
-        ];
-        return $data;
+        return $this->formHelper->getFormDataErrors($form);
     }
 
     public function deleteUser($user)
     {
         $this->em->remove($user);
         $this->em->flush();
-    }
-
-    private function getErrorsFromForm(FormInterface $form)
-    {
-        $errors = array();
-        foreach ($form->getErrors() as $error) {
-            $errors[] = $error->getMessage();
-        }
-        foreach ($form->all() as $childForm) {
-            if ($childForm instanceof FormInterface) {
-                if ($childErrors = $this->getErrorsFromForm($childForm)) {
-                    $errors[$childForm->getName()] = $childErrors;
-                }
-            }
-        }
-        return $errors;
-    }
-
-    /**
-     * @param $type
-     * @param null $data
-     * @param array $options
-     * @return \Symfony\Component\Form\FormInterface
-     */
-    public function createForm($type, $data = null, array $options = array())
-    {
-        return $this->formFactory->create($type, $data, $options);
     }
 }
